requirements.txt
python-dotenv
openai
pinecone
pypdf
tqdm
sentence-transformers
torch
